Moved all weather related data classes to HomeViewModel
- Keep WeatherFeature lightweight with composable classes only
- We will eventually want to move the network, data, and model classes to the correct modules in core, but at least now we have a better seperation of data and composables
