+++ feature/home/src/main/kotlin/com/ekhonavigator/feature/home/WeatherFeature.kt
@@ -1,8 +1,6 @@
 package com.ekhonavigator.feature.home
 
 import android.Manifest
-import android.annotation.SuppressLint
-import android.content.Context
 import android.content.pm.PackageManager
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -33,7 +31,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -46,63 +43,16 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.ekhonavigator.core.designsystem.icon.EkhoIcons
-import com.google.android.gms.location.LocationServices
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.tasks.await
-import kotlinx.coroutines.withContext
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import java.net.HttpURLConnection
-import java.net.URL
-import java.time.LocalDateTime
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
-import kotlin.math.roundToInt
-
-private data class WeatherUiState(
-    val isLoading: Boolean = true,
-    val currentTemperature: String = "--°F",
-    val conditionLabel: String = "Loading weather...",
-    val locationLabel: String = "Getting your location",
-    val windLabel: String = "-- mph",
-    val humidityLabel: String = "--%",
-    val highLowLabel: String = "H --°  L --°",
-    val hourlyForecast: List<HourlyForecastUi> = emptyList(),
-    val errorMessage: String? = null,
-)
-
-private data class HourlyForecastUi(
-    val timeLabel: String,
-    val temperatureLabel: String,
-    val conditionLabel: String,
-)
-
-@Serializable
-private data class OpenMeteoResponse(
-    @SerialName("current") val current: CurrentWeather,
-    @SerialName("hourly") val hourly: HourlyWeather,
-)
-
-@Serializable
-private data class CurrentWeather(
-    @SerialName("temperature_2m") val temperature2m: Double,
-    @SerialName("relative_humidity_2m") val relativeHumidity2m: Int,
-    @SerialName("wind_speed_10m") val windSpeed10m: Double,
-    @SerialName("weather_code") val weatherCode: Int,
-)
-
-@Serializable
-private data class HourlyWeather(
-    val time: List<String>,
-    @SerialName("temperature_2m") val temperature2m: List<Double>,
-    @SerialName("weather_code") val weatherCode: List<Int>,
-)
 
 @Composable
-internal fun WeatherSection(modifier: Modifier = Modifier) {
+internal fun WeatherSection(
+    viewModel: HomeViewModel,
+    modifier: Modifier = Modifier,
+) {
     val context = LocalContext.current
+    val weatherState by viewModel.weatherState.collectAsStateWithLifecycle()
 
     var hasLocationPermission by remember {
         mutableStateOf(
@@ -122,36 +72,24 @@ internal fun WeatherSection(modifier: Modifier = Modifier) {
     ) { permissions ->
         hasLocationPermission = permissions[Manifest.permission.ACCESS_FINE_LOCATION] == true ||
             permissions[Manifest.permission.ACCESS_COARSE_LOCATION] == true
+
+        if (hasLocationPermission) {
+            viewModel.loadWeather(context)
+        } else {
+            viewModel.onPermissionDenied()
+        }
     }
 
-    LaunchedEffect(Unit) {
-        if (!hasLocationPermission) {
+    LaunchedEffect(hasLocationPermission) {
+        if (hasLocationPermission) {
+            viewModel.loadWeather(context)
+        } else {
             permissionLauncher.launch(
                 arrayOf(
                     Manifest.permission.ACCESS_FINE_LOCATION,
                     Manifest.permission.ACCESS_COARSE_LOCATION,
                 ),
             )
-        }
-    }
-
-    val weatherState by produceState(
-        initialValue = WeatherUiState(
-            isLoading = hasLocationPermission,
-            conditionLabel = if (hasLocationPermission) "Loading weather..." else "Location permission needed",
-            locationLabel = if (hasLocationPermission) "Getting your location" else "Tap Allow to see your local weather",
-        ),
-        key1 = hasLocationPermission,
-    ) {
-        value = if (!hasLocationPermission) {
-            WeatherUiState(
-                isLoading = false,
-                conditionLabel = "Location permission needed",
-                locationLabel = "Tap Allow to see your local weather",
-                errorMessage = "Allow location to load current weather.",
-            )
-        } else {
-            loadWeatherState(context)
         }
     }
 
@@ -366,125 +304,4 @@ private fun FullDayForecastDialog(
             }
         },
     )
-}
-
-private suspend fun loadWeatherState(context: Context): WeatherUiState {
-    return try {
-        val location = getCurrentLocation(context)
-            ?: return WeatherUiState(
-                isLoading = false,
-                conditionLabel = "Location unavailable",
-                locationLabel = "Turn on location and try again",
-                errorMessage = "Couldn't get current location.",
-            )
-
-        val response = fetchWeather(location.latitude, location.longitude)
-        val todayForecast = buildTodayForecast(response)
-
-        WeatherUiState(
-            isLoading = false,
-            currentTemperature = fahrenheitLabel(response.current.temperature2m),
-            conditionLabel = weatherCodeToLabel(response.current.weatherCode),
-            locationLabel = "📍 ${formatLocationLabel(location.latitude, location.longitude)}",
-            windLabel = "${response.current.windSpeed10m.roundToInt()} mph",
-            humidityLabel = "${response.current.relativeHumidity2m}%",
-            highLowLabel = buildHighLowLabel(todayForecast),
-            hourlyForecast = todayForecast,
-        )
-    } catch (error: Exception) {
-        WeatherUiState(
-            isLoading = false,
-            conditionLabel = "Weather unavailable",
-            locationLabel = "Unable to reach weather service",
-            errorMessage = error.message ?: "Unknown error",
-        )
-    }
-}
-
-@SuppressLint("MissingPermission")
-private suspend fun getCurrentLocation(context: Context): android.location.Location? {
-    val fusedLocationClient = LocationServices.getFusedLocationProviderClient(context)
-    return try {
-        fusedLocationClient.lastLocation.await()
-    } catch (_: Exception) {
-        null
-    }
-}
-
-private suspend fun fetchWeather(latitude: Double, longitude: Double): OpenMeteoResponse {
-    return withContext(Dispatchers.IO) {
-        val url = URL(
-            "https://api.open-meteo.com/v1/forecast" +
-                "?latitude=$latitude" +
-                "&longitude=$longitude" +
-                "&current=temperature_2m,relative_humidity_2m,wind_speed_10m,weather_code" +
-                "&hourly=temperature_2m,weather_code" +
-                "&temperature_unit=fahrenheit" +
-                "&wind_speed_unit=mph" +
-                "&timezone=auto"
-        )
-
-        val connection = (url.openConnection() as HttpURLConnection).apply {
-            requestMethod = "GET"
-            connectTimeout = 15000
-            readTimeout = 15000
-        }
-
-        try {
-            val body = connection.inputStream.bufferedReader().use { it.readText() }
-            Json { ignoreUnknownKeys = true }.decodeFromString<OpenMeteoResponse>(body)
-        } finally {
-            connection.disconnect()
-        }
-    }
-}
-
-private fun buildTodayForecast(response: OpenMeteoResponse): List<HourlyForecastUi> {
-    val deviceZone = ZoneId.systemDefault()
-    val today = LocalDateTime.now(deviceZone).toLocalDate()
-    val formatter = DateTimeFormatter.ofPattern("h:mm a")
-
-    return response.hourly.time.indices.mapNotNull { index ->
-        val dateTime = runCatching {
-            LocalDateTime.parse(response.hourly.time[index])
-        }.getOrNull() ?: return@mapNotNull null
-
-        if (dateTime.toLocalDate() != today) {
-            return@mapNotNull null
-        }
-
-        HourlyForecastUi(
-            timeLabel = dateTime.format(formatter),
-            temperatureLabel = fahrenheitLabel(response.hourly.temperature2m[index]),
-            conditionLabel = weatherCodeToLabel(response.hourly.weatherCode[index]),
-        )
-    }
-}
-
-private fun buildHighLowLabel(hours: List<HourlyForecastUi>): String {
-    val temps = hours.mapNotNull { hour ->
-        hour.temperatureLabel.removeSuffix("°F").toIntOrNull()
-    }
-    if (temps.isEmpty()) {
-        return "H --°  L --°"
-    }
-    return "H ${temps.maxOrNull()}°  L ${temps.minOrNull()}°"
-}
-
-private fun fahrenheitLabel(value: Double): String = "${value.roundToInt()}°F"
-
-private fun formatLocationLabel(latitude: Double, longitude: Double): String {
-    return "${"%.3f".format(latitude)}, ${"%.3f".format(longitude)}"
-}
-
-private fun weatherCodeToLabel(code: Int): String = when (code) {
-    0 -> "Clear sky"
-    1, 2 -> "Partly cloudy"
-    3 -> "Overcast"
-    45, 48 -> "Foggy"
-    51, 53, 55, 56, 57 -> "Drizzle"
-    61, 63, 65, 66, 67, 80, 81, 82 -> "Rain"
-    71, 73, 75, 77, 85, 86 -> "Snow"
-    95, 96, 99 -> "Thunderstorm"
-    else -> "Weather update"
 }
